Work-in-progress fixes.
