+++ programl/models/ggnn/BUILD
@@ -41,6 +41,7 @@ py_library(
         "//programl/models:batch_results",
         "//programl/models:model",
         "//programl/proto:programl_py",
+        "//programl/util/py:progress",
         "//third_party/py/absl",
         "//third_party/py/numpy",
         "//third_party/py/torch",
