+++ programl/models/lstm/BUILD
@@ -25,6 +25,7 @@ py_library(
         "//programl/models:batch_data",
         "//programl/models:model",
         "//programl/proto:programl_py",
+        "//programl/util/py:progress",
         "//third_party/py/absl",
         "//third_party/py/numpy",
     ],
