+++ tests/benchmarks/BUILD
@@ -31,23 +31,24 @@ py_test(
     ],
 )
 
-#py_test(
-#    name = "benchmark_dataflow_ggnn",
-#    timeout = "long",
-#    srcs = ["benchmark_dataflow_ggnn.py"],
-#    deps = [
-#        "//programl/models/ggnn",
-#        "//programl/proto:programl_py",
-#        "//programl/util/py:progress",
-#        "//tests/plugins:llvm_program_graph",
-#        "//tests/plugins:llvm_reachability_features",
-#        "//tasks/dataflow:ggnn_batch_builder",
-#        "//tasks/dataflow:graph_loader",
-#        "//third_party/py/absl",
-#        "//third_party/py/tqdm",
-#    ],
-#)
-#
+py_test(
+    name = "benchmark_dataflow_ggnn",
+    timeout = "long",
+    srcs = ["benchmark_dataflow_ggnn.py"],
+    deps = [
+        "//programl/models/ggnn",
+        "//programl/proto:programl_py",
+        "//programl/util/py:progress",
+        "//programl/util/py:threaded_iterator",
+        "//tasks/dataflow:ggnn_batch_builder",
+        "//tasks/dataflow:graph_loader",
+        "//tests/plugins:llvm_program_graph",
+        "//tests/plugins:llvm_reachability_features",
+        "//third_party/py/absl",
+        "//third_party/py/tqdm",
+    ],
+)
+
 #py_test(
 #    name = "benchmark_dataflow_lstm",
 #    timeout = "long",
@@ -56,6 +57,7 @@ py_test(
 #        "//programl/models/lstm",
 #        "//programl/proto:programl_py",
 #        "//programl/util/py:progress",
+#        "//programl/util/py:threaded_iterator",
 #        "//tests/plugins:llvm_program_graph",
 #        "//tests/plugins:llvm_reachability_features",
 #        "//tasks/dataflow:graph_loader",
